+++ Libs/AceGUI-3.0/widgets/AceGUIWidget-ColorPicker.lua
@@ -143,7 +143,12 @@ local function Constructor()
 	colorSwatch.background = texture
 	texture:SetWidth(16)
 	texture:SetHeight(16)
-	texture:SetColorTexture(1, 1, 1)
+	if texture.SetColorTexture then
+		texture:SetColorTexture(1, 1, 1)
+	else
+		texture:SetTexture(1, 1, 1)
+		texture:SetVertexColor(1, 1, 1, 1)
+	end
 	texture:SetPoint("CENTER", colorSwatch)
 	texture:Show()